airflow_coder/requirements.txt
SQLAlchemy
sqlalchemy-redshift
redshift_connector 
requests
psycopg2-binary
python-dotenv
pandas
python-binance

